requirements.txt
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==3.0
cycler==0.10.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
nltk==3.5
numpy==1.19.1
pandas==1.0.5
pdfminer.six==20200726
Pillow==7.2.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
scikit-learn==0.23.1
scipy==1.5.2
six==1.15.0
sklearn==0.0
sortedcontainers==2.2.2
threadpoolctl==2.1.0
tqdm==4.48.0
Werkzeug==1.0.1
wincertstore==0.2
wordcloud==1.7.0
gunicorn
